+++ src/backend/resolvers/mutations/organization/obm/service/index.js
@@ -1,0 +1,7 @@
+const createObmService = require('./create')
+const updateObmService = require('./update')
+
+module.exports = {
+  createObmService,
+  updateObmService,
+}