+++ src/backend/resolvers/queries/organizations/obm/index.js
@@ -1,0 +1,9 @@
+const obmOrganizations = require('./obm')
+const obmServices = require('./service')
+const obmServicesCategories = require('./serviceCategory')
+
+module.exports = {
+  obmServices,
+  obmOrganizations,
+  obmServicesCategories,
+}