+++ src/backend/resolvers/queries/organizations/obm/service.js
@@ -1,0 +1,9 @@
+const obmServices = async (
+  parent,
+  args,
+  { pulseCoreDb },
+) => pulseCoreDb.collection('obm.services')
+  .find()
+  .toArray()
+
+module.exports = obmServices