+++ src/backend/resolvers/queries/organizations/obm/serviceCategory.js
@@ -1,0 +1,9 @@
+const obmServicesCategories = async (
+  parent,
+  args,
+  { pulseCoreDb },
+) => pulseCoreDb.collection('obm.services.categories')
+  .find()
+  .toArray()
+
+module.exports = obmServicesCategories