+++ src/backend/typeDefs/mutations/index.js
@@ -84,6 +84,10 @@ const mutationType = gql`
 
     createObmOrganization(input: CreateObmOrganizationInput!): CreateObmOrganizationPayload!
     updateObmOrganization(input: UpdateObmOrganizationInput!): UpdateObmOrganizationPayload!
+    createObmService(input: CreateObmServiceInput!): CreateObmServicePayload!
+    updateObmService(input: UpdateObmServiceInput!): UpdateObmServicePayload!
+    createObmServiceCategory(input: CreateObmServiceCategoryInput!): CreateObmServiceCategoryPayload!
+    updateObmServiceCategory(input: UpdateObmServiceCategoryInput!): UpdateObmServiceCategoryPayload!
 
     upsertOrganizationMeta(input: UpsertOrganizationMetaInput!): [UpsertOrganizationMetaPayload]