+++ src/backend/typeDefs/mutations/organization/obm/index.js
@@ -1,7 +1,12 @@
 const createObmOrganizationTypeDefs = require('./create')
 const updateObmOrganizationTypeDefs = require('./update')
 
+const obmServiceTypeDefs = require('./service')
+const obmServiceCategoryTypeDefs = require('./serviceCategory')
+
 module.exports = [
   createObmOrganizationTypeDefs,
   updateObmOrganizationTypeDefs,
+  ...obmServiceTypeDefs,
+  ...obmServiceCategoryTypeDefs,
 ]