+++ src/backend/typeDefs/mutations/organization/obm/service/create.js
@@ -1,0 +1,14 @@
+const { gql } = require('apollo-server-express')
+
+const createObmServiceTypeDefs = gql`
+  input CreateObmServiceInput {
+    name: String!
+  }
+
+  type CreateObmServicePayload {
+    _id: ID!
+    name: String!
+  }
+`
+
+module.exports = createObmServiceTypeDefs