+++ src/backend/typeDefs/mutations/organization/obm/service/index.js
@@ -1,0 +1,7 @@
+const createObmServiceTypeDefs = require('./create')
+const updateObmServiceTypeDefs = require('./update')
+
+module.exports = [
+  createObmServiceTypeDefs,
+  updateObmServiceTypeDefs,
+]