+++ src/backend/typeDefs/mutations/organization/obm/service/update.js
@@ -1,0 +1,15 @@
+const { gql } = require('apollo-server-express')
+
+const updateObmServiceTypeDefs = gql`
+  input UpdateObmServiceInput {
+    _id: ID!
+    name: String!
+  }
+
+  type UpdateObmServicePayload {
+    _id: ID!
+    name: String!
+  }
+`
+
+module.exports = updateObmServiceTypeDefs